backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-docx
python-pptx
google-generativeai
python-dotenv